Move `model_dump` override to `BaseModel` so that all models are guranteed serializable
